Plot hull points as polygon

+++ include/usv_sim_2d/visualisation.hpp
@@ -16,7 +16,7 @@ public:
 private:
     sf::RenderWindow window_;
 
-    const int zoom_ = 100;
+    const float zoom_ = 100;
 
     sf::Vector2f transform_coord_system(const Eigen::Vector3d &position, float zoom, sf::Vector2f offset);
     void draw_grid(int rows, int cols);
